refactor(MatchStatus): MatStatus는 결국 enum으로 관리되는 상수이므로 constant로 이동
