+++ src/main/java/dto/PlayerWinning.java
@@ -1,6 +1,6 @@
 package dto;
 
-import model.MatchStatus;
+import constant.MatchStatus;
 
 public record PlayerWinning(PlayerName name, MatchStatus matchStatus) {
 }
