+++ src/main/java/view/OutputView.java
@@ -4,7 +4,7 @@ package view;
 import java.util.List;
 import java.util.Map.Entry;
 import model.DealerWinning;
-import model.MatchStatus;
+import constant.MatchStatus;
 import model.PlayersWinning;
 import dto.Card;
 import dto.ParticipantWinning;
