+++ src/test/java/service/TestBlackJackService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 import model.BlackJackDeck;
 import model.Dealer;
-import model.MatchStatus;
+import constant.MatchStatus;
 import model.Player;
 import dto.PlayerName;
 import model.Players;
